% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{every_news}
\alias{every_news}
\title{Everything}
\usage{
every_news(q = NULL, results = 100, sources = NULL, domains = NULL,
  exclude_domains = NULL, since = NULL, until = NULL,
  language = NULL, sort = NULL)
}
\arguments{
\item{q}{Keywords or a phrase to search for.}

\item{results}{Number of results to return, on free plan this is often \code{100}. Set to \code{Inf} to return everything.}

\item{sources}{A \code{vector} of identifiers for the news sources or blogs you want headlines from. 
Use \code{\link{news_sources}} to locate these. 
Note: you can't mix this argument with the \code{country} or \code{category} argument.}

\item{domains, exclude_domains}{A \code{vector} of domains to restrict the search to, or exclude, i.e.: \code{bbc.co.uk}.}

\item{since, until}{\code{Date} or \code{POSIX} indicating oldest and newest article allowed.}

\item{language}{Find sources that display news in a specific language. 
Possible options: \code{ar}, \code{de}, \code{en}, \code{es}, \code{fr}, \code{he}, 
\code{it}, \code{nl}, \code{no}, \code{pt}, \code{ru}, \code{se}, \code{ud}, \code{zh}. Default: all languages.}

\item{sort}{The order to sort the articles in. Possible options: \code{relevancy}, \code{popularity}, \code{publishedAt}.}
}
\description{
Search through all articles.
}
\examples{
\dontrun{
 every_news("Trump")
}

}
