% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_sources}
\alias{get_sources}
\title{get_sources}
\usage{
get_sources(category = "", language = "", country = "", apiKey = NULL,
  parse = TRUE)
}
\arguments{
\item{category}{Name of news category.}

\item{language}{Name of language must be one of en, de, or fr. Defaults
to all.}

\item{country}{Name of country must be one of au, de, gb, in, it, us. Defaults
to all.}

\item{apiKey}{Character string API token. Default is to grab it from user R environ.}

\item{parse}{Logical indicating whether to parse response object to data frame.}
}
\value{
Data frame or nested list.
}
\description{
Returns news sources with meta data.
}
\examples{
\dontrun{
df <- get_sources()
}
}
