% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{news_sources}
\alias{news_sources}
\title{News Sources}
\usage{
news_sources(country = NULL, category = NULL, language = NULL)
}
\arguments{
\item{country}{The 2-letter ISO 3166-1 code of the country you want to get headlines for.}

\item{category}{The category you want to get headlines for. 
Possible options: \code{business}, \code{entertainment}, \code{general}, \code{health}, \code{science}, \code{sports}, \code{technology}. 
Note: you can't mix this argument with the \code{sources} argument.}

\item{language}{Find sources that display news in a specific language. 
Possible options: \code{ar}, \code{de}, \code{en}, \code{es}, \code{fr}, \code{he}, 
\code{it}, \code{nl}, \code{no}, \code{pt}, \code{ru}, \code{se}, \code{ud}, \code{zh}. Default: all languages.}
}
\description{
Get the list of available sources.
}
\examples{
\dontrun{
 news_sources()
}

}
