% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{newsapi_key}
\alias{newsapi_key}
\title{Setup}
\usage{
newsapi_key(key)
}
\arguments{
\item{key}{Your API key, freely available at \url{https://newsapi.org}.}
}
\description{
Setup your session, all subsequent calls will be done using the API key.
}
\examples{
\dontrun{
newsapi_key("xXXxxXxXxXXx")  
}

}
