% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{parse_articles}
\alias{parse_articles}
\title{parse_articles}
\usage{
parse_articles(x)
}
\arguments{
\item{x}{Response object returned by get_articles.}
}
\value{
Data frame
}
\description{
Converts response object from get articles to data frame
}
