% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{parse_sources}
\alias{parse_sources}
\title{parse_sources}
\usage{
parse_sources(x)
}
\arguments{
\item{x}{Response object from get_sources.}
}
\value{
Data frame
}
\description{
Parses sources response object into data frame
}
