% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{top_headlines}
\alias{top_headlines}
\title{Top Headlines}
\usage{
top_headlines(q = NULL, results = 100, country = NULL,
  category = NULL, sources = NULL)
}
\arguments{
\item{q}{Keywords or a phrase to search for.}

\item{results}{Number of results to return, on free plan this is often \code{100}. Set to \code{Inf} to return everything.}

\item{country}{The 2-letter ISO 3166-1 code of the country you want to get headlines for.}

\item{category}{The category you want to get headlines for. 
Possible options: \code{business}, \code{entertainment}, \code{general}, \code{health}, \code{science}, \code{sports}, \code{technology}. 
Note: you can't mix this argument with the \code{sources} argument.}

\item{sources}{A \code{vector} of identifiers for the news sources or blogs you want headlines from. 
Use \code{\link{news_sources}} to locate these. 
Note: you can't mix this argument with the \code{country} or \code{category} argument.}
}
\description{
Returns top headlines.
}
\examples{
\dontrun{
top_headlines("Trump")
}

}
